refactor: Moved YOLOv4 to yolo submodule
