+++ holocron/models/detection/__init__.py
@@ -1,2 +1,1 @@
 from .yolo import *
-from .new_yolo import *
